requirements.txt
appdirs==1.4.4
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
# data-wizard==1.3.0
distlib==0.3.2
Django==2.2.24
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
html-json-forms==1.1.1
idna==3.2
install==1.3.4
itertable==2.0.0
Markdown==3.3.4
mysqlclient==2.0.3
natural-keys==1.6.0
pbr==5.6.0
pip==21.1.2
pipenv==2021.5.29
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
setuptools==57.0.0
six==1.16.0
sqlparse==0.4.1
stevedore==3.3.0
urllib3==1.26.6
virtualenv==20.4.7
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wheel==0.36.2
xlrd==2.0.1